Podcast: Pre-caching - A podcast need not provide an image
It is not mandatory that a podcast feed provide an image, and it is not
helpful to attempt a refresh every day just because one is absent. This
is not an error condition.

So treat the default image just as if it were a 'real' image, caching
for around 90 days.

The image data will be updated when the user browses into the feed, so
any future change will be picked up within a reasonable time period.