Fix scan progress reporting - 'raw' path names need to be decoded
The 'name' given to a newly created scan progress report often, but not
always, includes a path element. That path element consists of raw
bytes.

When the path contains non-ASCII characters, its display in the "Media
Scan Details" of the Web UI will be corrupted, because it is subjected
to a subsequent utf-8 encoding process when the web page is rendered.

This contrasts with a path given in the 'info' to a progress update
report, which is properly handled.

This change corrects matters by decoding the name before entry into the
DB, and specifying that the relevant DB column has utf-8 semantics
applied to it.

It replicates the logic already applied to an 'info' progress update.