Move noisy cuesheet log messages to debug level
Parsing cuesheets is noisy if a cuesheet doesn't have "REM END" commands.
I get the feeling that "REM END" is some non-standard command that not all
applications are required to use or understand. I can't find information
on "REM END" or which applications that use it.

So, logging error messages if "REM END" is missing seems wrong.

For me these error messages are extremely noisy because I have used
CD ripper Sound Juicer to rip many of my CDs and Sound Juicer embeds
cuesheets in FLAC files it writes which look like this:

FILE "01 The Song.flac" FLAC
  TRACK 01 AUDIO
    INDEX 00 19404
  TRACK 02 AUDIO
    INDEX 00 11982264
  ...

When scanning this cuesheet there will be error messages logged for every
track in the cuesheet.

If a ripped CD has N tracks which are saved to separate files (each with an
embedded cuesheet) there will be N * (N + N - 1) error messages when scanning
these files.
N=10 -> 190 error messages
N=20 -> 780 error messages