Fix foobar 1.6.6+ sending Icy-Metadata on any request
This happens when using proxied stream for a UPnPBridge player like foobar 1.6.6 and this is weird.

In proxied mode, foobar is told to GET stream from LMS <IP:9000>/stream.mp3. When receiving a *stream* with icy-metadata, the foobar spontaneously makes a request to the server to grab some information. That request contains a Icy-Metadata header LMS sends a normal response but because of that header, it also tries to send icy data and crashes (sort of). I know it's a foobar specific issue at this point, but it's a general safeguard, like it has already been put in otherplaces of S:W:Player.pm to *not* send icy-metadata when this is not a proper stream.