calculate seekdata in HTTP::open() is only timeOffset is provided
Like for File::open, there are cases where only the time offset of seek data is provided (e.g. when using a PlayList, Jump command with a seekdata, user wants a time not a bytes offset). We then must calculate, if possible, such offset right before sending the HTTP request. This did not happen when seeking as StreamingController::JumpToTime does the calculation. This worked with File::open because it does calculate the seekdata during open