Fix #547 - Don't add album of same name if adding tracks while renaming the existing tracks' album title
As we did add new tracks before updating modified tracks, we'd first add new tracks with the new album title, resulting in a new album. When we then updated the existing tracks we didn't assign the existing tracks to that new album, but simply modified the existing album title. This resulted in two albums of the same name, one with the old, one with the new tracks.

By changing the scan order (delete, update, add, instead of delete, add, update) we hopefully fix this without introducing new side-effects.