Fix #789 - Fix player slimproto state initialization
Change #699 changed the way how we initialize re-connecting devices. Previously the SlimProto state would have been initialized as a side-effect of calling `$client->stop`. With #699 we no longer do so all the time. Which resulted in SlimProto state to be uninitialized under certain circumstances.

This change makes sure we initialize the SlimProto state whenever we receive a new SlimProto message and the state is undefined.