Fix client playback start on reconnect
This is a fix to the changes introduced by Player restart proposal #699,
https://github.com/Logitech/slimserver/pull/699.

A client that disappears/disconnects for a few seconds/minutes before
reconnecting may start playing immediately it reconnects, for no good reason.

This change ensures that a client's playing state is recorded as soon as the
client disappears/disconnects. The playing state is used to determine if the
client should restart playback when it connects to LMS, so it is essential
that it is recorded as soon as the client is disconnected.

At present, the playing state (recorded in preference 'playingAtPowerOff') is
not recorded until LMS starts to 'forget' the disconnected client. This will
take place 5 minutes after 'slimproto_close', via a timer that calls
'forget_disconnected_client'.

But if the client reconnects in the interim, '_hello_handler' will kill the
relevant timer, and the true playing state is not recorded. It may, in fact,
reflect a condition that existed some days/weeks in the past.