Fix installation of select plugin updates only
This change corrects the behaviour of the Plugin management settings page.

At present, an available plugin update will be installed whenever settings are
saved, without regard to the user's choice in the matter.

The settings page provides the user with a checkbox option that should
determine whether or not the update is installed. But the state of this option
is not being properly determined.

The problem lies with the use of the condition 'exists $params->{"install:$plugin"}'.
This condition will be set true when a plugin has been selected for first time
installation. But it is also set true while the plugin remains installed.

The condition 'exists $params->{"update:$plugin"}' indicates that the user has
chosen to install an update.

The current logic assumes that only one of the conditions
'exists $params->{"install:$plugin"}' and 'exists $params->{"update:$plugin"}'
will ever be true at the same time. Were that the case, the code would cater
for both first time installation and update installation correctly. But the
assumption is incorrect. 'exists $params->{"install:$plugin"}' is also true
whenever a plugin update is available.

The consequence is that plugin updates are installed without regard to the
user's choice.

This change deals with the issue by distinguishing a first time installation
from the installation of a subsequent update, and using the correct logic for
each case.