Treat multiple comments as multi-line, i.e. join with "\n" in place of "/"
This change modifies the way multiple comments are combined together.

Some tagging schemes allow multiple comments. LMS expects a single
comment, so they need to be combined.

With this change, comments are combined with a newline separator, in
place of the existing " / ". They then effectively become multi-line comments.

The change also eliminates the removal of an 'eng' prefix to a comment. It has
been in the code since at least 2005, and appear to serve no purpose. It may
have once been needed to remove an 'eng' language tag that is found in
ID3 comments. But these tags are already removed by LMS during the scanning
process.