Tweak display of multi-line and lyrics info items in web interface
This change adjusts the display of comment and lyrics text in the
Default & Classic web interfaces. Its intention is to make longer
texts more readable by keeping the content together in a block.

LMS tags both comment and lyrics info items with the 'wrap' attribute, and
this is how they are identified.

LMS does not tag any other info items in this way. But if it, or a plugin,
does, then this slight change in behaviour will also apply to them.