Populate the release year ('year') attribute for remote tracks in the player queue.
Many client players apps, including the Default skin, Material skin and Squeeze Ctrl, now display the release year for tracks in the player queue as well as for the current track. For remote tracks, this attribute can be obtained from the track's metadata. However, the LMS code fails to pick it up correctly and all remote tracks in the player queue end up with a value of '0' in the 'year' attribute.  With this change, remote tracks for which the 'year' attribute is supplied in the metadata will have the value populated in the player queue. This currently includes all Radio Paradise tracks and, with an upcoming change to the Qobuz plugin, will also apply to tracks added to the queue from the plugin. It will also apply to tracks from any other subscription-based streaming services for which the 'year' attribute is populated. The changes have been thoroughly tested with remote radio streams including Radio Paradise, as well as with Qobuz, using multiple clients, including the LMS skins, Material Skin, Squeezer, Squeeze Ctrl and Orange (Open) Squeeze.