Don't start an artwork rescan before the database is really ready.
When a plugin registered a new image resolution to be pre-cached, that would immediately trigger an artwork scan. As this can happen during initialization, the scan could get interrupted, or abort because the database hasn't fully been initialized. Wait a few seconds before running that scan.

Fix the web UI for this preference while we are at it. Make sure the description isn't considered a regex, or we might start adding a name over and over again. And make that description a clickable label for the option.