Correction to Album listing (Default Skin) Track favourites link
See https://forums.slimdevices.com/forum/user-forums/logitech-media-server/1681581-track-favourites-in-default-skin

The problem is that $songinfo (in Slim/Web/XMLBrowser.pm) always contains the album header info in the album view, so it should not be used to create a track favourite.
I'm looking at $item instead, if it's a track (type==audio).

I've tried to be as light-touch as possible in this change, but:
1. Has it really been wrong effectively for ever?
2. Is this the best way of addressing the problem?