Sometimes setting a music or playlist folder might lead the server into an infinite recursion.
We would stop queueing scan requests before we store the playlist folder - which then would immediately trigger a rescan. If at the same time the music folder was invalid, this could somehow lead to an infinite recursion.

Keep a small, sort lived cache of recently scanned folders and skip them if scanned recently.