When evaluating "New Tracks", rely on the `added` timestamp only.
Use the oldest value. This means that the first track added to the library would define when all of the album was added. This can be considered wrong. But the opposite would be equally wrong. There's no absolute truth. We take this deliberate decision, knowing some will consider it wrong.