Make `TRACKARTIST` a fix part of the unified artists browse list.
This reverts PR #1186 in large parts, but adds some more consequent use of `Slim::Schema::Contributor->unifiedArtistsListRoles` instead. I believe the motivation for #1186 was to include custom roles in search result sets. This is one place where this PR does things differently than before #1186.

`TRACKARTIST` had been in the unified artists list for > 10 years. Removing it or making it optional IMHO was our mistake. But I believe we only did so in order to use other methods compiling role lists. This change will hopefully do the best of both...