Invalidate cached XML browser sub-feeds on quantity mismatch
If a client does a query for a small amount of data followed by a query
for a larger amount of data, the second query returned partially invalid
results since it returned the data of the first query, padded with
empty items. Fix this by re-querying data in that case.

Fixes #1298

Signed-off-by: Danny Baumann <dannybaumann@web.de>