Fix #1214 Remove duplicate call to `_getNextTrack()`
This must have been an accidental left-over in https://github.com/LMS-Community/slimserver/pull/985/files: while a condition was added for this call, the old unconditional call was not removed. This lead to the situation that `_getNextTrack()` would be called twice, putting a player into bad state, from which it wouldn't recover without a restart.

I've had this change in place for several months, and didn't encounter any issue.

Signed-off-by: Michael Herger <michael@herger.net>