Remove MySQL support
I didn't remove all related code. But all which dealt with starting our own server instance, hints about it etc. If somebody wants to pick up the pieces and bring MySQL support up to speed, it should be relatively easy to re-enable it. But we don't want to encourage people to do so.

Signed-off-by: Michael Herger <michael@herger.net>