Fix: BUG 17980 - Allow item removal even if playlist has only 1 item
At present there is no way to clear the item from the current playlist
when the playlist only contains that one item.

This change restores the ability to remove that item from the playlist
by adding the "Remove from playlist" context menu item, even when the
playlist only that 1 item.

This option had historically been present, but was removed in commit 71e36d4
(Unnecessary items in context menu for current-playlist tracks, 2012-02-22).

Signed-off-by: Martin Williams <martinr.williams@gmail.com>