Provide a default TimeZone to SqueezeOS - Changes following Code review
Renames module from 'Slim/Web/SqueezeosTimezone' to 'Slim/Web/Time'.

Changes HTTP endpoint from '/tz' to '/time/tz'.

Tidies up TimeZone validation checks to minimize the number of checks to be
made, and curtails unnecessary code comments.

Code simplification - sends all HTTP responses through a single helper
function: '_sendHTTPresponse'.

Replaces inappropriate HTTP 404 (Not found) response codes with:
  - where there is an API problem - 500 (Internal server error)
  - where the TimeZone string fails validation - 204 (No content)

Caches the TimeZone obtained from the API call for the duration of LMS's
run time, in a module global variable. This includes TimeZones that have
failed validation (i.e. recorded as ''), because the result is unlikely to
change if the API is called again. But API failures are excluded from this,
on the assumption that these are likely to be transient in nature.

Signed-off-by: Martin Williams <martinr.williams@gmail.com>