#1416 Fix picking up metadata for locally imported remote streams
Instead of adding more confusion to `Slim::Schema::objectForUrl()` add a new method `libraryObjectForUrl()` which would fall back to the former for local tracks, or remote tracks which have not been imported. This basically emulates the behaviour of PR #1403 without changing `objectForUrl`'s behaviour for the dozens or hundreds of other callers.

* revert PR #1403
* add `libraryObjectForUrl()` to look up remote tracks in local library
* fall back to previous implementation for other tracks
* replace calling `objectForUrl()` with calling `libraryObjectForUrl()` in the few instances where we're only interested in a track's metadata

Signed-off-by: Michael Herger <michael@herger.net>