Signed-off-by: Michael OReilly <oreillymj@gmail.com>
Reload the playlist frame after a drop event to rebuild the playback control urls.

After a drop event, the play, up, down and delete icons at the right side of the playlist frame have urls which are out of synch with the frame content.
Clicking on an item without reloading the frame results in unpredictable playback etc.
The easiest fix is to reload the frame, although you could re-write the urls from each item between the moved items.
The frame normally reloads after 30s anyway, so this is a bit of a edge case in terms of behaviour.