Define the `playcount` as the average of the tracks' play counts
Using the sum would prioritize albums with many tracks over albums with few tracks. By using the average we level this difference out.

OTOH this will discriminate albums from which the user consistently skips a few odd tracks, or never does listen to the end, as it might play for too long to be listened to in one go...

Signed-off-by: Michael Herger <michael@herger.net>